kvm: Fix build issues
This patch fixes build issues related to the kvm vLinux builds.

+++ scripts/linux/build_module.sh
@@ -108,7 +108,7 @@ function build_libfdt()
 
 function build_kvmtool()
 {
-	echo "== Build KVMTOOL =="
+	echo "== Build kvmtool =="
 	(
 	cd $KVMTOOL_PATH
 	make ARCH=$ARCH CROSS_COMPILE=$CROSS_COMPILE lkvm-static
@@ -450,6 +450,7 @@ function build_linux()
 			apply_modcfg linux my_defconfig d_sto.cfg
 		fi
 		if [ "xyes" = "x${BUILD_KVM}" ]; then
+			apply_modcfg linux my_defconfig e_mod.cfg
 			apply_modcfg linux my_defconfig e_kvm.cfg
 		fi
 		make ARCH=$ARCH CROSS_COMPILE=$CROSS_COMPILE mrproper
@@ -476,17 +477,25 @@ function build_linux()
 		echo "Error: Failed to build Linux"
 		exit 1
 	fi
+	make ARCH=$ARCH CROSS_COMPILE=$CROSS_COMPILE \
+		O=$LINUX_BUILD/ -j`nproc` modules
 	${CROSS_COMPILE}objcopy \
 		--only-keep-debug $LINUX_BUILD/vmlinux \
 		$LINUX_BUILD/kernel.sym
 	if [ "xyes" = "x${BUILD_VLINUX}" ]; then
+		if [ ! -f $LINUX_BUILD/arch/$ARCH/kvm/kvm.ko ]; then
+			echo "Error: Failed to build vLinux"
+			exit 1
+		fi
 		mkdir -p $DESTDIR
-		#(
-		#cd $LINUX_BUILD
-		#make ARCH=$ARCH INSTALL_MOD_PATH=$DESTDIR modules_install
-		#make ARCH=$ARCH INSTALL_PATH=$DESTDIR install
-		#)
-		cp -f $LINUX_BUILD/arch/$ARCH/kvm/kvm.ko ${APPDIR}/
+		make ARCH=$ARCH CROSS_COMPILE=$CROSS_COMPILE \
+			O=$LINUX_BUILD/ \
+			INSTALL_MOD_STRIP=1 \
+			INSTALL_MOD_PATH=$DESTDIR modules_install
+		#cp -f $LINUX_BUILD/arch/$ARCH/kvm/kvm.ko ${APPDIR}/
+		#make ARCH=$ARCH CROSS_COMPILE=$CROSS_COMPILE \
+		#	O=$LINUX_BUILD/ \
+		#	INSTALL_PATH=$DESTDIR install
 		cp -f $LINUX_BUILD/arch/$ARCH/boot/Image ${APPDIR}/
 	fi
 	)
